refactor(utils/send-sign-in-email): minimize sign in email link sends
